+++ test/inject-loggedUser-koa1-api-key.spec.ts
@@ -2,7 +2,6 @@
 import Koa1 from "koa1";
 import nock from 'nock';
 import chai, { expect } from 'chai';
-import { RWAPIMicroservice } from 'main';
 import type Logger from "bunyan";
 import type { LogLevel } from "bunyan";
 import bunyan from "bunyan";
@@ -19,6 +18,7 @@ import {
 } from "./utils/mocks";
 import { BootstrapArguments } from "../src/types";
 import { mockCloudWatchLogRequest, mockCloudWatchSetupRequestsSequence } from "../src/test-mocks";
+import { RWAPIMicroservice } from "../src/main";
 
 chai.should();
 chai.use(ChaiHttp);