+++ test/inject-loggedUser-koa2-api-key-microservice-token.spec.ts
@@ -2,7 +2,6 @@ import type Koa from "koa";
 import Koa2 from "koa";
 import nock from 'nock';
 import chai, { expect } from 'chai';
-import { RWAPIMicroservice } from 'main';
 import type Logger from "bunyan";
 import type { LogLevel } from "bunyan";
 import bunyan from "bunyan";
@@ -18,6 +17,7 @@ import {
     mockCloudWatchSetupRequestsSequence
 } from "../src/test-mocks";
 import { BootstrapArguments } from "../src/types";
+import { RWAPIMicroservice } from "../src/main";
 
 chai.should();
 chai.use(ChaiHttp);