+++ test/inject-loggedUser-koa2-api-key-user-token.spec.ts
@@ -1,7 +1,6 @@
 import Koa2 from "koa";
 import nock from 'nock';
 import chai, { expect } from 'chai';
-import { RWAPIMicroservice } from 'main';
 import type Logger from "bunyan";
 import type { LogLevel } from "bunyan";
 import bunyan from "bunyan";
@@ -15,6 +14,7 @@ import ChaiHttp from 'chai-http';
 import { mockValidateRequestWithApiKeyAndUserToken } from "./utils/mocks";
 import { BootstrapArguments } from "../src/types";
 import { mockCloudWatchLogRequest, mockCloudWatchSetupRequestsSequence } from "../src/test-mocks";
+import { RWAPIMicroservice } from "../src/main";
 
 chai.should();
 chai.use(ChaiHttp);