+++ test/inject-loggedUser-koa2-microservice-token.spec.ts
@@ -1,7 +1,6 @@
 import Koa2 from "koa";
 import nock from 'nock';
 import chai, { expect } from 'chai';
-import { RWAPIMicroservice } from 'main';
 import type Logger from "bunyan";
 import type { LogLevel } from "bunyan";
 import bunyan from "bunyan";
@@ -15,6 +14,7 @@ import ChaiHttp from 'chai-http';
 import { mockValidateRequestWithUserToken } from "./utils/mocks";
 import { BootstrapArguments } from "../src/types";
 import { mockCloudWatchLogRequest, mockCloudWatchSetupRequestsSequence } from "../src/test-mocks";
+import { RWAPIMicroservice } from "../src/main";
 
 chai.should();
 chai.use(ChaiHttp);
@@ -51,7 +51,7 @@ describe('Injecting logged user data - Koa v2.x with MICROSERVICE token', () => 
             gatewayURL: 'https://controltower.dev',
             microserviceToken: constants.MICROSERVICE_TOKEN,
             fastlyEnabled: false,
-            requireAPIKey: false,
+            requireAPIKey: true,
             awsRegion: 'us-east-1',
             awsCloudWatchLogStreamName: 'test',
         };
@@ -105,7 +105,7 @@ describe('Injecting logged user data - Koa v2.x with MICROSERVICE token', () => 
             gatewayURL: 'https://controltower.dev',
             microserviceToken: constants.MICROSERVICE_TOKEN,
             fastlyEnabled: false,
-            requireAPIKey: false,
+            requireAPIKey: true,
             awsRegion: 'us-east-1',
             awsCloudWatchLogStreamName: 'test',
         };
@@ -160,7 +160,7 @@ describe('Injecting logged user data - Koa v2.x with MICROSERVICE token', () => 
             gatewayURL: 'https://controltower.dev',
             microserviceToken: constants.MICROSERVICE_TOKEN,
             fastlyEnabled: false,
-            requireAPIKey: false,
+            requireAPIKey: true,
             awsRegion: 'us-east-1',
             awsCloudWatchLogStreamName: 'test',
         };
@@ -218,7 +218,7 @@ describe('Injecting logged user data - Koa v2.x with MICROSERVICE token', () => 
             gatewayURL: 'https://controltower.dev',
             microserviceToken: constants.MICROSERVICE_TOKEN,
             fastlyEnabled: false,
-            requireAPIKey: false,
+            requireAPIKey: true,
             awsRegion: 'us-east-1',
             awsCloudWatchLogStreamName: 'test',
         };
@@ -276,7 +276,7 @@ describe('Injecting logged user data - Koa v2.x with MICROSERVICE token', () => 
             gatewayURL: 'https://controltower.dev',
             microserviceToken: constants.MICROSERVICE_TOKEN,
             fastlyEnabled: false,
-            requireAPIKey: false,
+            requireAPIKey: true,
             awsRegion: 'us-east-1',
             awsCloudWatchLogStreamName: 'test',
         };